backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.33.0
cryptography==41.0.2
Django==4.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jwt==1.3.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
